Remove etcd from the slurmctld operator (#13)
* Remove etcd from the slurmctld

Etcd is currently used to signal to the slurmd nodes that they
can start. However, an alternative approach has been determined in
which systemd services will be used with a start exit hook to call
back into charm code to perform this function. As such, etcd should
be removed to simplify the install footprint and function of a
Charmed Slurm deployment.

Since etcd is being removed, the tls certificates are no longer
necessary so this patch removes those also.

Signed-off-by: Billy Olsen <billy.olsen@canonical.com>

* chore(deps): Bump version of ops to 2.*

Signed-off-by: Jason C. Nucciarone <jason.nucciarone@canonical.com>

* chore(lint): Automatically fix lint issues with ruff

Signed-off-by: Jason C. Nucciarone <jason.nucciarone@canonical.com>

---------

Signed-off-by: Billy Olsen <billy.olsen@canonical.com>
Signed-off-by: Jason C. Nucciarone <jason.nucciarone@canonical.com>
Co-authored-by: Billy Olsen <billy.olsen@canonical.com>